+++ library/src/main/cpp/AndroidAudioFairy.cpp
@@ -2,15 +2,17 @@
 
 AndroidAudioFairy::AndroidAudioFairy(int sampling, int bit, int channel) {
     pthread_mutex_init(&mutex, NULL);
-    memset(emptyBuffer, 0, sizeof(emptyBuffer));
     this->sampling = sampling;
     this->bit = bit;
     this->channel = (SLuint32) channel;
     skip = 0;
+    buffered = false;
+    captureQueue = NULL;
     init_sl();
 }
 
 AndroidAudioFairy::~AndroidAudioFairy() {
+    endCapture();
     lock();
     if (sl.slBufQ) {
         (*sl.slBufQ)->Clear(sl.slBufQ);
@@ -169,7 +171,12 @@ int AndroidAudioFairy::init_sl2() {
         return -1;
     }
     res = (*sl.slPlayObj)->GetInterface(sl.slPlayObj, SL_IID_ANDROIDSIMPLEBUFFERQUEUE, &sl.slBufQ);
-    if (SL_RESULT_SUCCESS != res) {
+    return SL_RESULT_SUCCESS != res ? -1 : 0;
+}
+
+int AndroidAudioFairy::startPlaying() {
+    SLresult res;
+    if (!sl.slBufQ || !sl.slPlay) {
         return -1;
     }
     res = (*sl.slBufQ)->RegisterCallback(sl.slBufQ, callback, this);
@@ -177,14 +184,7 @@ int AndroidAudioFairy::init_sl2() {
         return -1;
     }
     res = (*sl.slPlay)->SetPlayState(sl.slPlay, SL_PLAYSTATE_PLAYING);
-    if (SL_RESULT_SUCCESS != res) {
-        return -1;
-    }
-    res = (*sl.slBufQ)->Enqueue(sl.slBufQ, emptyBuffer, sizeof(emptyBuffer));
-    if (SL_RESULT_SUCCESS != res) {
-        return -1;
-    }
-    return 0;
+    return SL_RESULT_SUCCESS != res ? -1 : 0;
 }
 
 void AndroidAudioFairy::lock() {
@@ -210,8 +210,14 @@ void AndroidAudioFairy::callback(SLAndroidSimpleBufferQueueItf bq, void *c) {
             for (int i = 0; i < copiedSize; i += context->skip + 1, c++) {
                 context->skipBuffer[c] = context->buffer[i];
             }
+            if (context->captureQueue) {
+                context->captureQueue->enqueue(context->skipBuffer, c * 2);
+            }
             (*bq)->Enqueue(bq, context->skipBuffer, c * 2);
         } else {
+            if (context->captureQueue) {
+                context->captureQueue->enqueue(context->buffer, copiedSize * 2);
+            }
             (*bq)->Enqueue(bq, context->buffer, (SLuint32) copiedSize * 2);
         }
     }