+++ test/src/main/java/com/suzukiplan/emulator/nes/test/CaptureVideoDialog.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 
-class CaptureDialog : DialogFragment() {
+class CaptureVideoDialog : DialogFragment() {
     private lateinit var capture: Bitmap
 
     fun show(manager: FragmentManager, capture: Bitmap) {
@@ -19,7 +19,7 @@ class CaptureDialog : DialogFragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater?.inflate(R.layout.dialog_capture, container)
+        val view = inflater?.inflate(R.layout.dialog_capture_video, container)
         view?.findViewById<ImageView>(R.id.capture_preview)?.setImageBitmap(capture)
         return view
     }